Make keymap an optional argument of modalka-define-key
With an optional keymap argument it is possible to also create prefix
commands in "normal" mode. Eg. like a mapping from `x s` to `C-x C-s`:

  (let* ((keymap (define-prefix-command 'modalka-Control-X-prefix)))
    (define-key modalka-mode-map (kbd "x") keymap)
    (modalka-define-key (kbd "s") (kbd "C-x C-s") keymap))

Note, that I am by no means an Emacs wizard, so it is quite possible
that this can be achieved much simpler or cleaner or is a total stupid
idea altogether.

Related to #25.